Initial Refactor
