+++ frontend-next/components/app/sidebar/profile.js
@@ -1,0 +1,7 @@
+export function Profile_Sidebar() {
+    return (
+        <div className="h-dvh">
+          <div className=" bg-white m-2 h-[98%]">Profile</div>
+        </div>
+    )
+}
